let admin see the running state of other users
The admin user can see the running state of other users under Settings
-> System -> Advanced -> Multiple users. If a user is "Active", then
they haven't been logged out/stopped yet.

If the user navigates back to the Multiple users page before the least
recently used users are shutdown, they can see an outdated list. So, we
listen for Intent.ACTION_SHUTDOWN broadcasts that are sent when a
user shuts down to make sure the state of the list is updated. See
com.android.server.am.UserController#finishUserStopping for where the
Intent is broadcasted.